+++ packages/orq-rc/src/models/operations/createdataset.ts
@@ -138,7 +138,7 @@ export const CreateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-09-14T08:33:02.935Z",
+    "2026-09-14T10:10:20.455Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {